requirements.txt
numpy
scikit-learn
matplotlib
tensorflow
Pillow
moviepy
ipython
